Further row implementation. No Tests yet.
