+++ tensorboard/webapp/plugins/npmi/store/npmi_reducers.ts
@@ -123,8 +123,7 @@ const reducer = createReducer(
       }
       return {
         ...state,
-        flaggedAnnotations: [...combinedSelectedAnnotations],
-        selectedAnnotations: [],
+        selectedAnnotations: [...combinedSelectedAnnotations],
       };
     }
   ),
