+++ tensorboard/webapp/plugins/npmi/views/annotations_list/BUILD
@@ -30,7 +30,7 @@ ng_module(
         "//tensorboard/webapp/plugins/npmi/util:filter_annotations",
         "//tensorboard/webapp/plugins/npmi/util:metric_type",
         "//tensorboard/webapp/plugins/npmi/util:sort_annotations",
-        "//tensorboard/webapp/plugins/npmi/views/annotations_list/annotations_list_row",
+        "//tensorboard/webapp/plugins/npmi/views/annotations_list/annotation",
         "//tensorboard/webapp/plugins/npmi/views/annotations_list/annotations_list_toolbar",
         "//tensorboard/webapp/plugins/npmi/views/annotations_list/header",
         "//tensorboard/webapp/plugins/npmi/views/annotations_list/legend",
