+++ tensorboard/webapp/plugins/npmi/views/annotations_list/annotation/annotation_component.ng.html
@@ -26,7 +26,7 @@
       svgIcon="visibility_off_24px"
     ></mat-icon>
   </div>
-  <div #annotation class="annotation-chart">
+  <div #chart class="annotation-chart">
     <svg style="width: 100%; height: {{runs.length * 30}}px"></svg>
   </div>
 </div>
