+++ tensorboard/webapp/plugins/npmi/views/annotations_list/annotation/annotation_component.scss
